Fix all clippy warnings and upgrade wasmtime 29→36 (LTS) for security
Clippy: resolved all warnings across 43 files in 13 crates. Fixes include:
- too_many_arguments suppressions where refactoring would be worse
- clamp() instead of .max().min() chains
- while-let loops instead of loop/if-let/break
- is_none_or() instead of map_or(true, ...)
- needless_range_loop allowances for parallel-array Block-STM
- complex type aliases (ScanResult, PrepareResult)
- dead_code allowances for serde-deserialized API response fields
- collapsible if-let/match patterns
- unwrap_or_default, direct iteration, borrowed expression fixes

Wasmtime: upgraded from v29.0.1 to v36.0.6 (LTS). Resolves all 5
security advisories (RUSTSEC-2026-0020, 0021, 2025-0046, 0118, 0006).
No API changes required — drop-in upgrade.

All 548 tests pass. Zero clippy warnings. Only remaining cargo-audit
issue is ring 0.16.20 (transitive via libp2p, low risk).

+++ crates/pichain-network/src/bridge.rs
@@ -290,7 +290,7 @@ impl CircuitBreaker {
     pub fn status(&self, current_height: u64) -> CircuitBreakerStatus {
         let rolling_volume = self.rolling_volume(current_height);
         let utilization_bps = if self.max_hourly_volume > 0 {
-            ((rolling_volume as u128).saturating_mul(10_000) / self.max_hourly_volume) as u64
+            (rolling_volume.saturating_mul(10_000) / self.max_hourly_volume) as u64
         } else {
             0
         };
